feat: Add migration and control service #36

+++ backend/src/main/java/ch/puzzle/pcts/dto/certificate/CertificateDto.java
@@ -1,4 +1,6 @@
 package ch.puzzle.pcts.dto.certificate;
 
-public record CertificateDto(Long id, String name, double points, String comment) {
+import java.math.BigDecimal;
+
+public record CertificateDto(Long id, String name, BigDecimal points, String comment) {
 }
